requirements.txt
setuptools
python-dotenv
google-generativeai
streamlit
pandas
numpy
biopython
sentence-transformers
faiss-cpu
transformers
torch
groq
ctgan
joblib
langchain
langchain-core
langchain-groq
langchain-community
requests
python-Levenshtein
fastapi
uvicorn
python-multipart
aiofiles
accelerate
chromadb
pypdf
tiktoken
sentencepiece
sentence_transformers
pinecone
PyPDF2
Pillow
tools
PyMuPDF
table_evaluator
matplotlib
plotly
seaborn
tensorflow 
keras 
rdkit
tf-keras
firebase-admin
pyrebase4